# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller._ResponseAllFortune:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        $ref: '#/definitions/model.AllModel'
        description: 数据
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  controller._ResponseAuspicious:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/model.CalendarModel'
        type: array
      msg:
        description: 提示信息
        type: string
    type: object
  controller._ResponseCalendar:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        $ref: '#/definitions/model.CalendarModel'
        description: 数据
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  controller._ResponseOneFortune:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        $ref: '#/definitions/model.OneModel'
        description: 数据
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  model.AllModel:
    properties:
      month:
        $ref: '#/definitions/model.OneModel'
        description: 本年运势
        type: object
      today:
        $ref: '#/definitions/model.OneModel'
        description: 今日运势
        type: object
      tomorrow:
        $ref: '#/definitions/model.OneModel'
        description: 明日运势
        type: object
      week:
        $ref: '#/definitions/model.OneModel'
        description: 本周运势
        type: object
      year:
        $ref: '#/definitions/model.OneModel'
        type: object
    type: object
  model.CalendarModel:
    properties:
      caishen:
        description: 财神
        type: string
      chong:
        description: 冲
        type: string
      date:
        description: 日期
        type: string
      days:
        description: 还有几天
        type: integer
      emonth:
        description: 英文月
        type: string
      eweek:
        description: 英文周
        type: string
      fushen:
        description: 福神
        type: string
      ji:
        description: 忌
        type: string
      jiri:
        description: 吉日
        type: string
      jishenyiqu:
        description: 吉神宜趋
        type: string
      nongli:
        description: 农历
        type: string
      sha:
        description: 煞
        type: string
      shengxiao:
        description: 生肖
        type: string
      star:
        description: 星座
        type: string
      suici:
        description: 岁次
        type: string
      taishen:
        description: 胎神
        type: string
      week:
        description: 中文周
        type: string
      wuxing:
        description: 五行
        type: string
      xiongshen:
        description: 凶神
        type: string
      xishen:
        description: 喜神
        type: string
      yangli:
        description: 阳历
        type: string
      yi:
        description: 宜
        type: string
      zhiri:
        description: 值日
        type: string
    type: object
  model.OneModel:
    properties:
      career:
        description: 工作运势
        type: string
      color:
        description: 幸运颜色
        type: string
      health:
        description: 健康运势
        type: string
      love:
        description: 爱情运势
        type: string
      money:
        description: 财运运势
        type: string
      number:
        description: 幸运数字
        type: string
      presummary:
        description: 概述
        type: string
      star:
        description: 贵人星座
        type: string
      summary:
        description: 综合运势
        type: string
    type: object
  model.ParamAuspiciousList:
    properties:
      dur:
        description: 时间范围(必填) week:1周, month:1月, month3:3月, year:1年
        example: month
        type: string
      name:
        description: 名称(必填)
        example: 结婚
        type: string
      page:
        description: 页数(选填)    默认:1
        example: 1
        type: integer
      size:
        description: 条数(选填)    默认:20
        example: 20
        type: integer
      type:
        description: 类型(必填) yi:宜,ji:忌
        example: yi
        type: string
      weekend:
        description: 是否只看周末(选填) 1:是，0:否，默认:0
        example: true
        type: boolean
    required:
    - dur
    - name
    - type
    type: object
  model.ParamStar:
    properties:
      date:
        description: 日期(必填)
        example: "2020-11-11"
        type: string
      star:
        description: 星座(必填)
        example: baiyang
        type: string
      type:
        description: 时间(选填) 默认获取5个时间信息
        type: string
    required:
    - date
    - star
    type: object
host: 192.168.3.180:8082
info:
  contact: {}
  description: 玄学宝典接口文档
  license: {}
  title: 玄学宝典
  version: "1.0"
paths:
  /calendar:
    get:
      consumes:
      - application/json
      description: 黄历查询
      parameters:
      - description: 日期：例如：2020-11-11
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "2000":
          description: 黄历详情
          schema:
            $ref: '#/definitions/controller._ResponseCalendar'
      summary: 黄历查询
      tags:
      - 黄历相关
  /calendar/auspicious:
    get:
      consumes:
      - application/json
      description: 择吉日查询
      parameters:
      - description: 时间范围(必填) week:1周, month:1月, month3:3月, year:1年
        example: month
        in: query
        name: dur
        required: true
        type: string
      - description: 名称(必填)
        example: 结婚
        in: query
        name: name
        required: true
        type: string
      - description: 页数(选填)    默认:1
        example: 1
        in: query
        name: page
        type: integer
      - description: 条数(选填)    默认:20
        example: 20
        in: query
        name: size
        type: integer
      - description: 类型(必填) yi:宜,ji:忌
        example: yi
        in: query
        name: type
        required: true
        type: string
      - description: 是否只看周末(选填) 1:是，0:否，默认:0
        example: true
        in: query
        name: weekend
        type: boolean
      produces:
      - application/json
      responses:
        "2000":
          description: 吉日列表
          schema:
            $ref: '#/definitions/controller._ResponseAuspicious'
      summary: 择吉日查询
      tags:
      - 黄历相关
  /fortune:
    get:
      consumes:
      - application/json
      description: 星座运势查询
      parameters:
      - description: 日期(必填)
        example: "2020-11-11"
        in: query
        name: date
        required: true
        type: string
      - description: 星座(必填)
        example: baiyang
        in: query
        name: star
        required: true
        type: string
      - description: 时间(选填) 默认获取5个时间信息
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "2001":
          description: 获取星座某时间段运势
          schema:
            $ref: '#/definitions/controller._ResponseOneFortune'
        "2002":
          description: 获取星座所有时间段运势
          schema:
            $ref: '#/definitions/controller._ResponseAllFortune'
      summary: 星座运势查询
      tags:
      - 星座相关
swagger: "2.0"